Move agent/bootstrap to top-level bootstrap
The bootstrap is pretty well encapsulated and the agent sub-package is
getting to be pretty god-like in size. This moves bootstrap on it's own.
